# apps/backend/internal/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  backend_api_http.InternalServerErrorResponse:
    properties:
      errors: {}
      message:
        type: string
      status:
        example: Internal Server Error
        type: string
    type: object
  backend_api_http.Response:
    properties:
      data: {}
      errors: {}
      message:
        type: string
      status:
        type: string
    type: object
...
  backend_internal_db_repository.CreateWatchlistParams:
    properties:
      ca:
        example: SP1Y5YSTAHZ88XYK1VPDH24GY0HPX5J4JECTMY4A1.velar-token
        type: string
      userId:
        type: string
    required:
    - ca
    type: object
...
  backend_internal_db_repository.Watchlist:
    properties:
      ca:
        example: SP1Y5YSTAHZ88XYK1VPDH24GY0HPX5J4JECTMY4A1.velar-token
        type: string
      createdAt:
        type: string
      id:
        type: string
      updatedAt:
        type: string
      userId:
        type: string
    required:
    - ca
    type: object
...
host: localhost:8080
info:
  contact:
    email: atomic.k.2739@gmail.com
    name: Atomic
    url: https://github.com/iatomic1
  description: API for Dexion
  title: DEXION API
  version: "1.0"
paths:
...
  /watchlist:
    get:
      description: Retrieve all watchlist entries for the authenticated user
      produces:
      - application/json
      responses:
        "200":
          description: User watchlist retrieved
          schema:
            allOf:
            - $ref: '#/definitions/backend_api_http.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/backend_internal_db_repository.Watchlist'
                  type: array
              type: object
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/backend_api_http.InternalServerErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Get user's watchlist
      tags:
      - Watchlist
    post:
      consumes:
      - application/json
      description: Add a contract address to user's watchlist
      parameters:
      - description: Watchlist data
        in: body
        name: WatchlistRequest
        required: true
        schema:
          $ref: '#/definitions/backend_internal_db_repository.CreateWatchlistParams'
      produces:
      - application/json
      responses:
        "201":
          description: Watchlist created successfully
          schema:
            allOf:
            - $ref: '#/definitions/backend_api_http.Response'
            - properties:
                data:
                  $ref: '#/definitions/backend_internal_db_repository.Watchlist'
              type: object
        "400":
          description: Invalid request data
          schema:
            additionalProperties:
              type: string
            type: object
        "409":
          description: Contract already in watchlist
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/backend_api_http.InternalServerErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Create a new watchlist entry
      tags:
      - Watchlist
  /watchlist/{id}:
    delete:
      description: Delete a specific watchlist entry by ID
      parameters:
      - description: Watchlist ID (UUID)
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: Watchlist entry deleted successfully
          schema:
            $ref: '#/definitions/backend_api_http.Response'
        "400":
          description: Invalid ID
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Watchlist entry not found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/backend_api_http.InternalServerErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Delete watchlist entry
      tags:
      - Watchlist
schemes:
- http
- https
securityDefinitions:
  AccessTokenBearer:
    description: AccessTokenBearer Authentication
    in: header
    name: Authorization
    type: apiKey
  Google OAuth2:
    authorizationUrl: https://accounts.google.com/o/oauth2/auth
    flow: accessCode
    scopes:
      email: Grants access to user's email
      profile: Grants access to user's basic profile info
    tokenUrl: https://oauth2.googleapis.com/token
    type: oauth2
  RefreshTokenBearer:
    description: RefreshTokenBearer Authentication
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"
...